Added all project models to admin site
All models are now on admin site, with basic editing of all properties working.
Many improvements to be made, but everything should be possible to construct, but it may take a lot of time to generate large amounts of test data.

Fixed all the model __str__ methods that were not working.

Added some test data using the admin panel to test everything was working.