Moved from states to api requests on pages
Passing states had the limitation of not working when a in user goes directly to a page as there is nothing to pass to the link or route object. Hence made projects page obtain its data using an API request.

Modified projects.jsx to add a component that risk evaluations are mapped to, but only the most recent risk evaluation is shown (this is not the proper way of showing the most recent evaluation but it works).