Connected machine learning! (kinda)
ML is now estimate project success chances for projects in the database and risk evaluation objects are being created with the success chance saved to them. These are shown on the project's react page.

Not all the inputs to the ML model are being fed in because they will take some processing to create/get into correct format but it should be easy enough, famous last words...